+++ arenatacticsserver/scenes/UI/characters/character_select_screen/UI/character_select_ui.gd
@@ -1,7 +1,6 @@
 extends Control
 class_name CharacterSelectUI
 
-var selected_class: ClassDefinition = null
 var selected_skin: String = "default"
 
 @onready var css_hex_grid: CSSHexagonGrid = $"%CSSHexagonGrid"
@@ -22,28 +21,27 @@ func select_class(_selected_class: ClassDefinition) -> void:
 	if _selected_class == null:
 		select_random_class()
 		return
-	anim_player.play_backwards("fade_in")
-	selected_class = _selected_class
-	#print_debug("Class selected : %s" % selected_class)
 	class_selected.emit(_selected_class)
-	await anim_player.animation_finished
-	fill_data()
-	anim_player.play("fade_in")
 
+func update_selected_class(_selected_class: ClassDefinition):
+	anim_player.play_backwards("fade_in")
+	await anim_player.animation_finished
+	fill_data(_selected_class)
+	anim_player.play("fade_in")
 
 ## Get a random key from keys() then use that
 ## If the random class is the currently selected one, we roll again, but no more than 5 times to avoid infinite loop in the case there is only one class.
-func select_random_class() -> void:
+func select_random_class(current_class: ClassDefinition = null) -> void:
 	var i: int = 0
 	var rando_class = ClassLoader.classes[ClassLoader.classes.keys().pick_random()]
-	while rando_class == selected_class:
+	while rando_class == current_class:
 		rando_class = ClassLoader.classes[ClassLoader.classes.keys().pick_random()]
 		i += 1
 		if i > 4:
 			break
 	select_class(rando_class)
 
-func fill_data() -> void:
+func fill_data(selected_class: ClassDefinition) -> void:
 	char_title.text = selected_class.title.to_pascal_case()
 	char_subtitle.text = selected_class.subtitle
 	char_description.text = selected_class.description