+++ arenatacticsserver/scenes/world/Tiles/fake_hex_tile.gd
@@ -18,19 +18,7 @@ func _ready():
 	hexagon_prism.material.transparency = 1
 	mesh_inst.mesh = hexagon_prism
 
-	# var cylinder: CylinderShape3D = CylinderShape3D.new()
-	# cylinder.height = 0.5
-	# cylinder.radius = 0.45
-	# collision.shape = cylinder
-	# collision.position.y = height / 2 - (cylinder.height / 2)
-
 	anim_player.play("fade_in")
 
 func set_vertical_position():
-	#print_debug("%d:%d : %f, %d" % [grid_pos_x, grid_pos_y, distance, height])
-	
-	#self.position.y += abs(islandism)
-	# floating island. If you remove 2m of height, you must up the y pos of 1
-
-	# all flat bottom
 	self.position.y = (height / 2)